# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  admin.CreateInstanceRequest:
    properties:
      cpu:
        type: integer
      disk:
        type: integer
      image:
        type: string
      instance_type:
        type: string
      memory:
        type: integer
      name:
        type: string
      provider:
        type: string
      userId:
        type: integer
    required:
    - image
    - name
    - provider
    type: object
...
  admin.UpdateInstanceTypePermissionsRequest:
    properties:
      minLevelForContainer:
        maximum: 5
        minimum: 1
        type: integer
      minLevelForVM:
        maximum: 5
        minimum: 1
        type: integer
    type: object
...
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
host: localhost:8888
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 一键虚拟化管理平台API接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: OneClickVirt API
  version: "1.0"
paths:
...
  /admin/dashboard:
    get:
      consumes:
      - application/json
      description: 获取管理员后台首页的统计数据
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取管理员仪表板
      tags:
      - 管理员管理
...
  /admin/instance-type-permissions:
    get:
      consumes:
      - application/json
      description: 管理员获取实例类型最低等级要求配置
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: object
              type: object
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取实例类型权限配置
      tags:
      - 管理员管理
    put:
      consumes:
      - application/json
      description: 管理员更新实例类型最低等级要求配置
      parameters:
      - description: 权限配置参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.UpdateInstanceTypePermissionsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 更新失败
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 更新实例类型权限配置
      tags:
      - 管理员管理
  /admin/instances:
    get:
      consumes:
      - application/json
      description: 管理员获取系统中所有实例的列表，支持分页和过滤
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: pageSize
        type: integer
      - description: 实例状态
        in: query
        name: status
        type: string
      - description: 提供商筛选
        in: query
        name: provider
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取实例列表
      tags:
      - 管理员管理
    post:
      consumes:
      - application/json
      description: 管理员创建新的虚拟化实例
      parameters:
      - description: 创建实例请求参数
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.CreateInstanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 服务器内部错误
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 创建实例
      tags:
      - 管理员管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"